+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/HardforkId.java
@@ -56,6 +56,7 @@ public interface HardforkId {
     THANOS,
     MAGNETO,
     MYSTIQUE,
-    SPIRAL
+    SPIRAL,
+    SINISTER,
   }
 }